Breaking the whole thing out into modules
This will be more readable for bigger applications,
but for an application as small as this is might not really make sense.

I also simplified the model a little bit by removing the client-side
tracking of the uid, and just parsed the server response and inserted
that item. No more optimistic updates on creates!